TASK-B/requirements.txt
torch
torchvision
efficientnet_pytorch
scikit-learn
tqdm
Pillow
numpy
